returns NaN instead of -infinity with vacuum/analyze
We were returning "-infinity" on check_last_{analyze|vacuum} when a
database never had such a maintenance. First, this was not the
appropriate value, then check_pgactivity and the dispatcher were dealing
wrongly with such a value, leading to recording big negative values in
pgfactory.wh_nagios.
We now returns "NaN" and teach nagios_dispatcher.pl to parse it so it is
recorded as such in the database. Next step is to teach the UI how to
deal with such numbers.