mecanism to avoid pgsql version check for offline services
Service archive_folder does not require a pgsql connexion to run.
Before this patch a pgsql version check was always done, leading to
error if check_pgactivity can not connect on a pgsql cluster before
checking the folder itself, which has no sense on a remote archiving
server as instance. Some more offline services should appears in futur,
so we pave the path for them as well.