Check for critical/warning requierement from the services themselves
Before this patch, there was one check in the global part of the code
that check for warning/critical args requierement with some services
exclusion when they were optionnal.
This patch put this check in the services themselves. It makes it easier
to check if the service require warning/critical from its code, and
allow for some more fine grained checks.
This patch some more checks about units of the thresholds in each
services and update the documentation.