Change perfdata and threashold on commit_ration
Before this patch, the service was reporting a 100% commit_ratio
activity even under no activity on a database, which is quite confusing.

This patch replace this commit_ratio perfdata and threshold with
rollback_ratio, wich is 0% under no activity. The new threshold
rollback_ratio will raise an alert if the ratio reported is grater than
than the threshold.

It does seems more natural.

On my way, rename the threshold "rollback" as "rollbacks".