Optimize btree_bloat queries
Use pg_statistics instead of pg_stats and use UNION to gather stats
from indexes fields and table fields from two distinct queries instead
of JOIN with an "OR" filter, leading to a very bad plan and minutes of
execution for database with thousands of indexes.

The service must now be executed assuperuser only.