Add service pg_dump_backup
This service allows to check each pg_dump backups for each database
existing in the cluster.

It is able to raise alarms if the delta size is too important between
last backups, if the last backup did not occurs or if the retention
policy is not respected.

Patch written by Ronan Dunklau, with some fix and additions from me.