Fix bad estimate bloat with btree_bloat (and table_bloat)
When the computed average length of a column is inferior to 1 (if ratio
of NULL values is important in the column), a wrong cast in the btree
bloat query report a data+padding length of 0, leading to wrong
estimated bloat.

This seems to impact the table_bloat query as well, but probably with a
lesser impact as it rely on ceil(). Fixed in there as well anyway.