Several fixes and additions for check_wal_receiver
First use pg_xlog_location_diff() to calculate the lag in feedback, as discussed
in Pull Request #80.

Second, make check_wal_receiver work with PostgreSQL 10 as pg_xlog_location_diff()
has been renamed to pg_wal_location_diff().

Also, move a comment to a more appropriate place, emit a message according if a
replication slot is in use or not.