Add new function silent_is_compat
The current is_compat function emits a warning if the PostgreSQL version
that not match the arguments.
The sequences_exhausted service uses this function to choose between two
code path according the PostgreSQL version. Now, the archiver service works
the same way.

Also patch the archiver service to use this new function.