Fix service temp_files to work with PostgreSQL 10 (#125)
* Fix service temp_files to work with PostgreSQL 10

The service temp_files monitors both live created temp files and get
its counters from pg_stat_database.
The live part is achieve through calls of the pg_ls_dir() function that
requires superuser privileges. As PostgreSQL 10 now brings the pg_monitor
role in order to help the user to monitor its db without superpowers.
This commit modifies the behaviour of temp_files in order to take
account of the live temp files if the user is superuser, or simply
skip this part if the monitoring user is a normal user.