Add a relation_size service to follow the size of a few tables, including aggregated tables.
By default, it lists all tables, indexes, materialized views.
--include 'tablename' lists one (ore more) relations and all its(their) indexes
   and (sub)partitions with their pg_relation_size
--agg aggregates a table and its partitions (both partitioning modes)
   including indexes and toast
--details shows vm,fsm,main,toast,idx of tables instead of just
   pg_relation_size (and aggregates the indexes)
--agg --details shows vm,fsm,main,toast,idx summed up for all partitions.

Warning and Critical (optional) are triggered by size (compared to each
returned line).

This service is fully functional but may need more work: the main query is a
bit tormented and may be simplified or tuned, the perl part is a bit naive.
The main query uses dynamic filtering to reduce the load on the server.
Complete query time is usually below the second on my laptop, even with
thousands of tables, but may need a few seconds in degenerated cases
(10'000 partitions); and much more time is spent in check_pgactivity
itself without --agg (but listing thousands of tables is pointless).