Add a checksum_errors service.
Note that --dbinclude and --dbexclude are not supported on purpose.  A checksum
failure is a critical problem, and giving a way to users to hide some of the
problems would do more harm than good.

The thresholds are only available for the unlikely event that failures
happened, all the underlying problems has been fixed, checked and double
checked, and the cluster is back to a working state BUT that did not required
to setup a new instance or reset the statistics.