(#325, last_maintnenance) Raise alerts on threshold change (#362)
last_maintenance: Raise alerts on threshold change

When there is no write, the last_analyze and last_vacuum raised no alert
when crossing the threshold. This is okay. But lowering the threshold
did not lead to an alert.

Add the thresholds in the MD5 stored in the status file to force
a new test when they change. That may lead to a new alert on
totally inactive server when check_pgactivity is upgraded.

Close #325